requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.32.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.8
langgraph==0.0.26
pydantic==2.6.3
openpyxl==3.1.2
pandas==2.2.1 